caishui-agent/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
pymysql==1.1.1
cryptography==42.0.8
redis==5.0.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=3.2,<4.0
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
openpyxl==3.1.3
pandas==2.2.2
aiofiles==23.2.1
sse-starlette==2.1.0
captcha==0.5.0
Pillow==10.3.0
loguru==0.7.2
